pyproject.toml
[project]
name = "refine_ea"
version = "0.1.1"
description = "Discriminate between alignment candidates using LLM based reasoning"
authors = [
    {name = "Paco Goze",email = "paco.goze@polytehnique.edu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    "torch>=2.0.0",
    "PyYAML>=6.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.41.0",
    "transformers (>=4.54.1,<5.0.0)",
    "huggingface-hub (>=0.34.3,<0.35.0)",
    "requests (>=2.32.4,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
